requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.5
pydantic>=2.8.2
langchain>=0.2.14
langchain-openai>=0.1.22
python-dotenv>=1.0.1
